Add support for UIT queues
Add support for submitting jobs to UIT queues

+++ molequeue/molequeueglobal.h
@@ -28,7 +28,7 @@ namespace MoleQueue
 {
 
 /// Type for various ids
-typedef quint32 IdType;
+typedef qint64 IdType;
 
 /// Constant value used for invalid ids
 const IdType InvalidId = std::numeric_limits<IdType>::max();
